[MISC] Move seqan3::views::trim_quality to seqan3/alphabet/views/
