+++ include/seqan3/range/views/all.hpp
@@ -16,6 +16,7 @@
 #include <seqan3/alphabet/views/rank_to.hpp>
 #include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/alphabet/views/to_rank.hpp>
+#include <seqan3/alphabet/views/trim_quality.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/enforce_random_access.hpp>
 #include <seqan3/range/views/get.hpp>
@@ -30,7 +31,6 @@
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to.hpp>
 #include <seqan3/range/views/translate.hpp>
-#include <seqan3/range/views/trim_quality.hpp>
 #include <seqan3/utility/views/convert.hpp>
 #include <seqan3/utility/views/deep.hpp>
