+++ test/snippet/range/views/trim_quality_dna5q.cpp
@@ -5,7 +5,7 @@
 #include <seqan3/alphabet/quality/aliases.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
 #include <seqan3/alphabet/views/to_char.hpp>
-#include <seqan3/range/views/trim_quality.hpp>
+#include <seqan3/alphabet/views/trim_quality.hpp>
 
 int main()
 {
